chore: move yeomanScaffold feature flag to core to avoid cycle dependency
