+++ packages/fx-core/src/plugins/resource/spfx/utils/utils.ts
@@ -78,7 +78,3 @@ export class Utils {
 export async function sleep(ms: number) {
   await new Promise((resolve) => setTimeout(resolve, ms));
 }
-
-export function yeomanScaffoldEnabled() {
-  return false;
-}
